Refactoring on to ServerConnection Class
I've done some refactoring to the ServerConnection Class so that is easier to all of us to connect to the Server, and execute the methods. I've also updated methods to callback methods so that we can recieve data from the server and use it. I've also created one example method for others to see how does it function, and how can we use them. I'll do now the testing of the Class